remove python and add tests

+++ scripts/aws_cleanup/index.js
@@ -2,13 +2,11 @@
 const moment = require('moment');
 const { ec2 } = require('@cumulus/aws-client/services');
 
-const shouldBeCleanedUp = (instanceObject) => {
+const shouldBeCleanedUp = (instanceObject, todayFunc) => {
   const timeoutKey = process.env.timeout_key || 'Rotate By';
   return instanceObject.Tags.reduce((ret, tag) => {
-    if (tag.Key === timeoutKey) {
-      if (moment(tag.value) < moment()) {
-        return true;
-      }
+    if (tag.Key === timeoutKey && moment(tag.Value) < todayFunc()) {
+      return true;
     }
     return ret;
   }, false);
@@ -17,20 +15,33 @@ const shouldBeCleanedUp = (instanceObject) => {
 const getInstancesToClean = async (client) => {
   const describeResponse = await client.describeInstances();
   const instances = describeResponse.Reservations.flatMap((reservation) => reservation.Instances);
-  return instances.filter(shouldBeCleanedUp).map((instance) => instance.InstanceId);
+  return instances.filter((instance) =>
+    shouldBeCleanedUp(instance, moment)
+  ).map((instance) => instance.InstanceId);
 };
 
 const handler = async () => {
   const client = ec2();
   const toClean = await getInstancesToClean(client);
-  const termination = await client.terminateInstances({
-    InstanceIds: toClean,
-  });
+  if (toClean.length) {
+    const termination = await client.terminateInstances({
+      InstanceIds: toClean,
+    });
+    return {
+      statusCode: 200,
+      message: `termination completed with response ${JSON.stringify(termination)}`,
+    };
+  }
   return {
     statusCode: 200,
-    message: `termination completed with response ${JSON.stringify(termination)}`,
+    message: `termination completed with no instances out of date`,
   };
 };
+
+module.exports = {
+  shouldBeCleanedUp,
+  handler
+}
 
 if (require.main === module) {
   handler(
